@@ -12,23 +12,23 @@
   <properties>
     <compiler-plugin.version>3.10.1</compiler-plugin.version>
     <maven.compiler.parameters>true</maven.compiler.parameters>
-    <maven.compiler.source>11</maven.compiler.source>
-    <maven.compiler.target>11</maven.compiler.target>
-    <org.apache.poi.version>5.2.2</org.apache.poi.version>
+    <maven.compiler.source>17</maven.compiler.source>
+    <maven.compiler.target>17</maven.compiler.target>
+    <org.apache.poi.version>5.2.4</org.apache.poi.version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
     <quarkus.platform.artifact-id>quarkus-bom</quarkus.platform.artifact-id>
     <quarkus.platform.group-id>io.quarkus.platform</quarkus.platform.group-id>
 
-    <quarkus.platform.version>2.11.2.Final</quarkus.platform.version>
+    <quarkus.platform.version>3.4.2</quarkus.platform.version>
     <camel-quarkus.platform.version>${quarkus.platform.version}</camel-quarkus.platform.version>
 
     <camel-quarkus.platform.group-id>${quarkus.platform.group-id}</camel-quarkus.platform.group-id>
     <camel-quarkus.platform.artifact-id>quarkus-camel-bom</camel-quarkus.platform.artifact-id>
 
     <surefire-plugin.version>3.0.0-M5</surefire-plugin.version>
-    <weld-se-core.version>3.1.9.Final</weld-se-core.version>
-    <weld-junit5.version>2.0.2.Final</weld-junit5.version>
+    <weld-se-core.version>5.1.2.Final</weld-se-core.version>
+    <weld-junit5.version>4.0.0.Final</weld-junit5.version>
     <hamcrest.version>2.2</hamcrest.version>
     <json.version>20220320</json.version>
     <microprofile-openapi-api.version>3.0</microprofile-openapi-api.version>
