@@ -2,9 +2,6 @@ package net.a.g.excel.grpc;
 
 import java.util.function.Function;
 
-import javax.inject.Inject;
-import javax.inject.Singleton;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -17,6 +14,8 @@ import com.google.protobuf.StringValue;
 
 import io.quarkus.grpc.GrpcService;
 import io.smallrye.mutiny.Multi;
+import jakarta.inject.Inject;
+import jakarta.inject.Singleton;
 import net.a.g.excel.engine.ExcelEngine;
 
 @GrpcService
