@@ -1,12 +1,11 @@
 package net.a.g.excel.rest;
 
-import javax.inject.Singleton;
-
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 
 import io.quarkus.jackson.ObjectMapperCustomizer;
+import jakarta.inject.Singleton;
 
 @Singleton
 public class ExcelJacksonConfiguration implements ObjectMapperCustomizer {
